pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.nitpick]
style = "https://raw.githubusercontent.com/wemake-services/wemake-python-styleguide/master/styles/nitpick-style-wemake.toml"


[tool.poetry]
name = "coursegenius-api"
description = "Course Genius API backend"
version = "0.1.0"
readme = "README.md"
authors = ["Towhidul Haque Roni"]


[tool.poetry.dependencies]
python = "3.10.11"
django = "4.2"
djangorestframework = "3.14.0"
django-split-settings = "^1.1"
django-axes = "^5.21"
django-csp = "^3.7"
django-health-check = "^3.16"
django-permissions-policy = "^4.15"
django-stubs-ext = "^0.8.0"
psycopg2-binary = "2.9"
gunicorn = "^20.0"
python-decouple = "^3.4"
bcrypt = "^3.2"
structlog = "^21.1"
drf-yasg = "^1.20.0"
requests = "^2.28.0"
django-rest-auth = "0.9.5"
django-filter = ">=2.4.0"
django-concurrency = "2.4"
celery = "5.2.7"
flower = "1.2.0"
redis = "4.5.4"
channels = "4.0.0"
daphne = "4.0.0"
channels-redis = "^4.1.0"
Werkzeug = "^2.2.3"
django-add-default-value = "^0.9.1"
django-deprecate-fields = "^0.1.1"
django-reset-migrations = "^0.4.0"
djangorestframework-simplejwt = "^4.8.0"
pycryptodome = "3.17"
python3-openid = "^3.2.0"
sentry-sdk = "^1.3.1"
django-cors-headers = "^3.8.0"
django-anymail = "^8.4"
Weasyprint = "^53.2"
pypdf2 = "3.0.1"
newrelic = "^6.8.1"
pip = "21.2.4"
django-redis = "^5.0.0"
pandas = "^1.3.2"
packaging="23.1"
factory-boy = "^3.2.0"
python-keycloak = "^0.26.1"
types-requests = "^2.25.6"
PyYAML = "^6.0"
websocket-client = "^1.5.1"


[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.2.2"
django-querycount = "^0.7.0"
django-migration-linter = "^3.0"
django-extra-checks = "^0.11"
django-coverage-plugin = "^2.0"
nplusone = "^1.0"

wemake-python-styleguide = "0.16.1"
flake8-pytest-style = "^1.5"
flake8-django = "^1.1"
flake8-logging-format = "^0.6"
flake8-bandit = "3.0.0"
nitpick = "^0.27"

pytest = "^7.3.1"
pytest-django = "4.5.2"
pytest-cov = "^2.12"
pytest-randomly = "^3.10"
pytest-deadfixtures = "^2.2"
pytest-testmon = "^1.1"
pytest-timeout = "^1.4"
django-test-migrations = "^1.1"
hypothesis = "^6.17"

mypy = "1.2.0"
django-stubs = "1.16.0"

sphinx = "^6.1.3"
sphinx-autodoc-typehints = "^1.23.0"
tomlkit = "^0.7"
doc8 = "^0.9"

yamllint = "^1.26"
safety = "^2.1.0"
dotenv-linter = "0.3.0"
polint = "^0.4"
dennis = "^0.9"
dump-env = "^1.3"
ipython = "8.12.0"
